+++ src/main/webapp/components/basic-search/basic-search.js
@@ -27,7 +27,7 @@ import {
 } from './basic-search-helper'
 import FacetedDropdown from '../faceted-dropdown'
 import { Location } from '../../location'
-import { makeDefaultSearchGeo } from '../../query-builder/filter'
+import { makeDefaultSearchGeo } from '../query-builder/filter'
 import { useApolloFallback } from '../../react-hooks'
 import SortOrder from '../sort-order/sort-order'
 import SourcesSelect from '../sources-select'
