+++ src/main/webapp/components/query-builder/filter-inputs/boolean-filter.tsx
