+++ src/main/webapp/components/query-builder/filter/filter-group.tsx
