+++ src/main/webapp/components/query-builder/filter/operator.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import useAnchorEl from '../../react-hooks/use-anchor-el'
+import useAnchorEl from '../../../react-hooks/use-anchor-el'
 import Button from '@material-ui/core/Button'
 import Popover from '@material-ui/core/Popover'
 import MenuItem from '@material-ui/core/MenuItem'
