+++ src/main/webapp/components/query-builder/filter/sample-attribute-definitions.tsx
