+++ src/main/webapp/components/query-builder/query-builder.tsx
@@ -6,7 +6,7 @@ import MenuItem from '@material-ui/core/MenuItem'
 import Box from '@material-ui/core/Box'
 import TextField from '@material-ui/core/TextField'
 import { getIn, setIn } from 'immutable'
-import useAnchorEl from '../react-hooks/use-anchor-el'
+import useAnchorEl from '../../react-hooks/use-anchor-el'
 import { defaultFilter } from './filter/filter-utils'
 import Filter from './filter/individual-filter'
 import QuerySettings from './query-settings'
