+++ src/main/webapp/components/query-builder/query-settings.tsx
@@ -8,11 +8,10 @@ import Collapse from '@material-ui/core/Collapse'
 import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown'
 import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp'
 import { QuerySettingsType } from './types'
-const ResultFormSelect = require('../components/user-settings/result-form-select')
-  .default
-const SourcesSelect = require('../components/sources-select').default
-const { FilterCard } = require('../components/containers')
-const SortOrder = require('../components/sort-order').default
+const ResultFormSelect = require('../user-settings/result-form-select').default
+const SourcesSelect = require('../sources-select').default
+const { FilterCard } = require('../containers')
+const SortOrder = require('../sort-order').default
 
 const getSorts = (sorts?: string[]) => {
   if (!sorts) return undefined
