+++ src/main/webapp/search-forms/route.tsx
@@ -5,9 +5,9 @@ import Snackbar from '@material-ui/core/Snackbar'
 import SnackbarContent from '@material-ui/core/SnackbarContent'
 import * as React from 'react'
 import { Fragment, useState } from 'react'
-import { defaultFilter } from '../query-builder/filter/filter-utils'
-import QueryBuilder from '../query-builder/query-builder'
-import { QueryType } from '../query-builder/types'
+import { defaultFilter } from '../components/query-builder/filter/filter-utils'
+import QueryBuilder from '../components/query-builder/query-builder'
+import { QueryType } from '../components/query-builder/types'
 import SearchFormEditor from './editor'
 import { SnackbarRetry as RetryNotification } from '../components/network-retry'
 import { ApolloError } from 'apollo-client/errors/ApolloError'
